working f and t bindings
