+++ crates/terminal/src/mappings/keys.rs
@@ -1,6 +1,6 @@
 /// The mappings defined in this file where created from reading the alacritty source
 use alacritty_terminal::term::TermMode;
-use gpui::keymap::Keystroke;
+use gpui::keymap_matcher::Keystroke;
 
 #[derive(Debug, PartialEq, Eq)]
 pub enum Modifiers {
@@ -273,6 +273,8 @@ fn modifier_code(keystroke: &Keystroke) -> u32 {
 
 #[cfg(test)]
 mod test {
+    use gpui::keymap_matcher::Keystroke;
+
     use super::*;
 
     #[test]
