+++ crates/workspace/src/workspace.rs
@@ -33,6 +33,7 @@ use gpui::{
     actions,
     elements::*,
     impl_actions, impl_internal_actions,
+    keymap_matcher::KeymapContext,
     platform::{CursorStyle, WindowOptions},
     AnyModelHandle, AnyViewHandle, AppContext, AsyncAppContext, Entity, ModelContext, ModelHandle,
     MouseButton, MutableAppContext, PathPromptOptions, PromptLevel, RenderContext, Task, View,
@@ -2588,7 +2589,7 @@ impl View for Workspace {
         }
     }
 
-    fn keymap_context(&self, _: &AppContext) -> gpui::keymap::Context {
+    fn keymap_context(&self, _: &AppContext) -> KeymapContext {
         let mut keymap = Self::default_keymap_context();
         if self.active_pane() == self.dock_pane() {
             keymap.set.insert("Dock".into());
